# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MotoHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/logs:
    get:
      operationId: listAdminLogs
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminLog" }
    post:
      operationId: createAdminLog
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAdminLogInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminLog" }
...
  /admin/reports/{id}:
    patch:
      operationId: updateAdminReport
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateReportInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
    delete:
      operationId: deleteAdminReport
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No content
  /reports:
    post:
      operationId: createReport
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReportInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
...
  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: published
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BlogPost" }
  /blog/posts/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogPost" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /admin/blog/posts:
    get:
      operationId: adminListBlogPosts
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BlogPost" }
    post:
      operationId: adminCreateBlogPost
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBlogPost" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogPost" }
  /admin/blog/posts/{id}:
    patch:
      operationId: adminUpdateBlogPost
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBlogPost" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogPost" }
    delete:
      operationId: adminDeleteBlogPost
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No content
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners for the homepage carousel
      parameters:
        - name: all
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Banner" }
  /admin/banners:
    get:
      operationId: adminListBanners
      tags: [banners]
      summary: List all banners (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Banner" }
    post:
      operationId: adminCreateBanner
      tags: [banners]
      summary: Create a new banner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBanner" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Banner" }
  /admin/banners/{id}:
    put:
      operationId: adminUpdateBanner
      tags: [banners]
      summary: Update a banner
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBanner" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Banner" }
    delete:
      operationId: adminDeleteBanner
      tags: [banners]
      summary: Delete a banner
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    AdminLog:
      type: object
      required: [id, adminId, adminName, action, target, details, createdAt]
      properties:
        id: { type: integer }
        adminId: { type: integer }
        adminName: { type: string }
        action: { type: string }
        target: { type: string }
        details: { type: string }
        createdAt: { type: string }
    CreateAdminLogInput:
      type: object
      required: [adminId, action]
      properties:
        adminId: { type: integer }
        adminName: { type: string }
        action: { type: string }
        target: { type: string }
        details: { type: string }
    Report:
      type: object
      required: [id, reporterId, targetType, targetId, reason, details, status, createdAt]
      properties:
        id: { type: integer }
        reporterId: { type: integer }
        targetType: { type: string }
        targetId: { type: integer }
        reason: { type: string }
        details: { type: string }
        status: { type: string, enum: [pending, resolved, dismissed] }
        createdAt: { type: string }
    CreateReportInput:
      type: object
      required: [reporterId, targetId, reason]
      properties:
        reporterId: { type: integer }
        targetType: { type: string }
        targetId: { type: integer }
        reason: { type: string }
        details: { type: string }
    UpdateReportInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, resolved, dismissed] }
...
    BlogPost:
      type: object
      required: [id, title, slug, content, excerpt, category, authorId, authorName, published, seoTitle, seoDescription, views, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        excerpt: { type: string }
        category: { type: string }
        authorId: { type: integer }
        authorName: { type: string }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"] }
        seoTitle: { type: string }
        seoDescription: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        views: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    CreateBlogPost:
      type: object
      required: [title, slug, content, authorId]
      properties:
        title: { type: string, minLength: 3 }
        slug: { type: string, minLength: 3 }
        content: { type: string }
        excerpt: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        authorId: { type: integer }
        authorName: { type: ["string", "null"] }
        published: { type: ["boolean", "null"] }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        coverImageUrl: { type: ["string", "null"] }
    UpdateBlogPost:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        excerpt: { type: string }
        category: { type: string }
        published: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
        coverImageUrl: { type: ["string", "null"] }
...
    Banner:
      type: object
      required: [id, title, subtitle, ctaText, ctaUrl, imageUrl, bgColor, order, active, durationSecs, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        subtitle: { type: string }
        ctaText: { type: string }
        ctaUrl: { type: string }
        imageUrl: { type: string }
        bgColor: { type: string }
        order: { type: integer }
        active: { type: boolean }
        startsAt: { type: ["string", "null"] }
        endsAt: { type: ["string", "null"] }
        durationSecs: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    CreateBanner:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        subtitle: { type: ["string", "null"] }
        ctaText: { type: ["string", "null"] }
        ctaUrl: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        bgColor: { type: ["string", "null"] }
        order: { type: ["integer", "null"] }
        active: { type: ["boolean", "null"] }
        startsAt: { type: ["string", "null"] }
        endsAt: { type: ["string", "null"] }
        durationSecs: { type: ["integer", "null"] }
    UpdateBanner:
      type: object
      properties:
        title: { type: string }
        subtitle: { type: string }
        ctaText: { type: string }
        ctaUrl: { type: string }
        imageUrl: { type: string }
        bgColor: { type: string }
        order: { type: integer }
        active: { type: boolean }
        startsAt: { type: ["string", "null"] }
        endsAt: { type: ["string", "null"] }
        durationSecs: { type: integer }
...